Added IntegerConvertible as base interface for EnumInteger and MaskValueSet
Added IntegerConvertible as base interface for EnumInteger and MaskValueSet. IntergerConvertible defines the intValue() method and marks classes that can be converted into ints and vice-versa.
Changed EnumInteger to explicitly wrap Enums as special case of IntegerConvertible
Changed MaskValueSet to extend IntegerConvertible and renamed method toInt to intValue in accordance to IntegerConvertible
Changed method calls to refelct this change